# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "PayableImportCsvRequest": {
        "description": "Arquivo CSV de contas a pagar.",
        "properties": {
          "file": {
            "description": "CSV UTF-8 com cabeçalho obrigatório:\ndescription,amount,status,expirationDate,paymentDate,supplierId\n",
            "format": "binary",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "type": "object"
      },
      "PayableImportationItemResponse": {
        "description": "Resultado do processamento de uma linha do CSV.",
        "properties": {
          "errorMessage": {
            "description": "Motivo da rejeição quando o item possui ERROR.",
            "type": "string"
          },
          "id": {
            "description": "Identificador do item.",
            "example": 10,
            "format": "int64",
            "type": "integer"
          },
          "payableId": {
            "description": "Conta criada quando o item possui SUCCESS.",
            "example": 25,
            "format": "int64",
            "type": "integer"
          },
          "payableImportationId": {
            "description": "Importação à qual o item pertence.",
            "example": 1,
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "description": "Resultado da linha.",
            "enum": [
              "SUCCESS",
              "ERROR"
            ],
            "example": "SUCCESS",
            "type": "string"
          }
        },
        "type": "object"
      },
      "PayableImportationResponse": {
        "description": "Importação assíncrona de contas a pagar.",
        "properties": {
          "createdAt": {
            "description": "Data e hora de criação.",
            "example": "2026-08-01T10:15:30",
            "type": "string"
          },
          "errorMessage": {
            "description": "Erro global quando a importação estiver em FAILED.",
            "type": "string"
          },
          "id": {
            "description": "Identificador da importação.",
            "example": 1,
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "description": "Estado do processamento assíncrono.",
            "enum": [
              "PENDING",
              "PROCESSING",
              "COMPLETED",
              "COMPLETED_WITH_ERRORS",
              "FAILED"
            ],
            "example": "PENDING",
            "type": "string"
          },
          "updatedAt": {
            "description": "Data e hora da última atualização.",
            "example": "2026-08-01T10:15:35",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PayableRequest": {
        "description": "Dados para criação ou atualização de uma conta a pagar.",
        "properties": {
          "amount": {
            "description": "Valor monetário não negativo, com até duas casas decimais.",
            "example": 1500.0,
            "minimum": 0,
            "type": "number"
          },
          "description": {
            "description": "Descrição da conta a pagar.",
            "example": "Pagamento de aluguel",
            "type": "string"
          },
          "expirationDate": {
            "description": "Data de vencimento no formato ISO-8601.",
            "example": "2026-08-10",
            "format": "date",
            "type": "string"
          },
          "paymentDate": {
            "description": "Data do pagamento, obrigatória somente para status PAGO.",
            "example": "2026-08-05",
            "format": "date",
            "type": "string"
          },
          "status": {
            "description": "Status da conta. Quando omitido, assume PENDENTE. PAGO exige paymentDate;\nos demais status não admitem data de pagamento.\n",
            "enum": [
              "PENDENTE",
              "PAGO",
              "CANCELADO"
            ],
            "example": "PENDENTE",
            "type": "string"
          },
          "supplierId": {
            "description": "Identificador de um fornecedor existente.",
            "example": 1,
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "amount",
          "description",
          "supplierId"
        ],
        "type": "object"
      },
      "PayableResponse": {
        "description": "Conta a pagar persistida.",
        "properties": {
          "amount": {
            "description": "Valor da conta.",
            "example": 1500.0,
            "type": "number"
          },
          "description": {
            "description": "Descrição da conta.",
            "example": "Pagamento de aluguel",
            "type": "string"
          },
          "expirationDate": {
            "description": "Data de vencimento.",
            "example": "2026-08-10",
            "format": "date",
            "type": "string"
          },
          "id": {
            "description": "Identificador da conta.",
            "example": 1,
            "format": "int64",
            "type": "integer"
          },
          "paymentDate": {
            "description": "Data do pagamento, quando a conta estiver paga.",
            "example": "2026-08-05",
            "format": "date",
            "type": "string"
          },
          "status": {
            "description": "Status atual da conta.",
            "enum": [
              "PENDENTE",
              "PAGO",
              "CANCELADO"
            ],
            "example": "PENDENTE",
            "type": "string"
          },
          "supplier": {
            "$ref": "#/components/schemas/SupplierResponse",
            "description": "Fornecedor vinculado."
          }
        },
        "type": "object"
      },
      "ProblemDetail": {
        "description": "Erro HTTP no formato RFC 9457.",
        "properties": {
          "detail": {
            "description": "Explicação específica do erro.",
            "example": "Dados inválidos",
            "type": "string"
          },
          "instance": {
            "description": "URI da requisição que originou o erro.",
            "type": "string"
          },
          "status": {
            "description": "Código HTTP.",
            "example": 400,
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "description": "Título correspondente ao status HTTP.",
            "example": "Bad Request",
            "type": "string"
          },
          "type": {
            "description": "Identificador do tipo do problema.",
            "example": "about:blank",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "description": "Credenciais configuradas por APP_SECURITY_USERNAME e APP_SECURITY_PASSWORD.",
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para gestão de fornecedores, contas a pagar, importações assíncronas em CSV\ne relatórios de pagamentos. Todos os recursos exigem autenticação HTTP Basic.\n",
    "title": "Payable Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/import/payable": {
      "get": {
        "description": "Retorna todas as importações de contas a pagar e seus estados atuais.",
        "operationId": "listPayableImportations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PayableImportationResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Importações existentes."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          }
        },
        "summary": "Listar importações",
        "tags": [
          "Importações"
        ]
      },
      "post": {
        "description": "Recebe um CSV UTF-8, salva o arquivo, cria uma importação PENDING e publica um\nevento no RabbitMQ. O 201 confirma a aceitação, não o término. O consumidor muda\no estado para PROCESSING e depois COMPLETED, COMPLETED_WITH_ERRORS ou FAILED.\n\nCabeçalho obrigatório:\n`description,amount,status,expirationDate,paymentDate,supplierId`.\nOs status aceitos são PENDENTE, PAGO e CANCELADO; datas usam `yyyy-MM-dd`.\nLinhas em branco são ignoradas e campos podem usar aspas duplas.\n",
        "operationId": "createPayableImportation",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PayableImportCsvRequest"
              }
            }
          },
          "description": "Arquivo CSV com extensão .csv.",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayableImportationResponse"
                }
              }
            },
            "description": "Importação aceita e aguardando processamento."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Dados inválidos",
                  "status": 400,
                  "title": "Bad Request",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Arquivo ausente, vazio ou sem extensão CSV."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "415": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "O tipo de mídia enviado não é suportado para a importação",
                  "status": 415,
                  "title": "Unsupported Media Type",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Requisição não usa multipart/form-data."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Não foi possível publicar o evento de importação",
                  "status": 500,
                  "title": "Internal Server Error",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Falha ao armazenar o arquivo ou publicar o evento."
          }
        },
        "summary": "Importar contas por CSV",
        "tags": [
          "Importações"
        ]
      }
    },
    "/import/payable/{id}": {
      "get": {
        "description": "Retorna uma importação pelo identificador para acompanhamento assíncrono.",
        "operationId": "findPayableImportationById",
        "parameters": [
          {
            "description": "Identificador da importação.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayableImportationResponse"
                }
              }
            },
            "description": "Importação encontrada."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Importação não encontrada."
          }
        },
        "summary": "Consultar importação",
        "tags": [
          "Importações"
        ]
      }
    },
    "/import/payable/{id}/items": {
      "get": {
        "description": "Retorna o resultado de cada linha processada. SUCCESS possui payableId;\nERROR possui errorMessage e não cria uma conta.\n",
        "operationId": "listPayableImportationItems",
        "parameters": [
          {
            "description": "Identificador da importação.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PayableImportationItemResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Itens processados da importação."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Importação não encontrada."
          }
        },
        "summary": "Listar itens da importação",
        "tags": [
          "Importações"
        ]
      }
    },
...
    "/payable/{id}": {
      "delete": {
        "description": "Exclui definitivamente uma conta a pagar existente.",
        "operationId": "deletePayable",
        "parameters": [
          {
            "description": "Identificador da conta a pagar.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Conta a pagar excluída."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conta a pagar não encontrada."
          }
        },
        "summary": "Excluir conta a pagar",
        "tags": [
          "Contas a pagar"
        ]
      },
      "get": {
        "description": "Retorna uma conta a pagar e o fornecedor associado.",
        "operationId": "findPayableById",
        "parameters": [
          {
            "description": "Identificador da conta a pagar.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayableResponse"
                }
              }
            },
            "description": "Conta a pagar encontrada."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conta a pagar não encontrada."
          }
        },
        "summary": "Consultar conta a pagar",
        "tags": [
          "Contas a pagar"
        ]
      },
      "put": {
        "description": "Substitui os dados da conta. O identificador da rota prevalece sobre eventual id do corpo.",
        "operationId": "updatePayable",
        "parameters": [
          {
            "description": "Identificador da conta a pagar.",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "amount": 1500.0,
                "description": "Pagamento de aluguel",
                "expirationDate": "2026-08-10",
                "status": "PENDENTE",
                "supplierId": 1
              },
              "schema": {
                "$ref": "#/components/schemas/PayableRequest"
              }
            }
          },
          "description": "Novos dados da conta a pagar.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayableResponse"
                }
              }
            },
            "description": "Conta a pagar atualizada."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Dados inválidos",
                  "status": 400,
                  "title": "Bad Request",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Dados violam uma regra da conta a pagar."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Credenciais de autenticação ausentes ou inválidas",
                  "status": 401,
                  "title": "Unauthorized",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Credenciais ausentes ou inválidas."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "detail": "Recurso não encontrado",
                  "status": 404,
                  "title": "Not Found",
                  "type": "about:blank"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conta ou fornecedor não encontrado."
          }
        },
        "summary": "Atualizar conta a pagar",
        "tags": [
          "Contas a pagar"
        ]
      }
    },
...
  },
  "security": [
    {
      "basicAuth": []
    }
  ],
  "servers": [
    {
      "description": "Servidor da aplicação",
      "url": "/management/rest"
    }
  ],
...
}